tera states refactor
