+++ packages/client/hmi-client/src/model-representation/service.ts
@@ -2,6 +2,7 @@
 import _, { isEmpty } from 'lodash';
 import { runDagreLayout, rerouteEdges } from '@/services/graph';
 import { MiraModel } from '@/model-representation/mira/mira-common';
+import { ModelPartItem } from '@/types/Model';
 import { extractNestedStratas } from '@/model-representation/petrinet/mira-petri';
 import type { Initial, Model, ModelParameter, State, RegNetVertex, Transition, Rate } from '@/types/Types';
 import { getModelType } from '@/services/model';
@@ -349,4 +350,39 @@ export function checkPetrinetAMR(amr: Model) {
 	});
 
 	return results;
+}
+
+export function createStateList(collapsedInitials, model) {
+	return Array.from(collapsedInitials.keys()).map((id) => {
+		const childTargets = collapsedInitials.get(id) ?? [];
+		const isParent = childTargets.length > 1;
+		const states = getStates(model); // could be states, vertices, and stocks type
+		const children = childTargets
+			.map((childTarget) => {
+				const s = states.find((state) => state.id === childTarget);
+				if (!s) return null;
+				return {
+					id: s.id,
+					name: s.name,
+					description: s.description,
+					grounding: s.grounding,
+					unitExpression: s.units?.expression
+				};
+			})
+			.filter(Boolean) as ModelPartItem[];
+
+		const baseState: any = states.find((s) => s.id === id);
+		const base: ModelPartItem =
+			isParent || !baseState
+				? { id }
+				: {
+						id,
+						name: baseState.name,
+						description: baseState.description,
+						grounding: baseState.grounding,
+						unitExpression: baseState.units?.expression
+					};
+
+		return { base, children, isParent };
+	});
 }
